basic-template/lambdas/financial_analyst/requirements.txt
openai
newspaper3k
yfinance
lxml_html_clean
phidata